+++ src/main/java/kitchenpos/order/domain/OrderRepository.java
@@ -1,6 +1,5 @@
-package kitchenpos.product.domain;
+package kitchenpos.order.domain;
 
-import kitchenpos.order.domain.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
