+++ src/test/java/kitchenpos/product/application/ProductServiceTest.java
@@ -1,5 +1,7 @@
 package kitchenpos.product.application;
 
+import kitchenpos.product.domain.Product;
+import kitchenpos.product.domain.ProductRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
