+++ src/test/java/kitchenpos/table/application/TableServiceTest.java
@@ -1,6 +1,6 @@
 package kitchenpos.table.application;
 
-import kitchenpos.product.domain.OrderRepository;
+import kitchenpos.order.domain.OrderRepository;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.OrderTableRepository;
 import kitchenpos.tableGroup.domain.TableGroup;
